+++ polars/polars-lazy/src/dsl/function_expr/mod.rs
@@ -22,6 +22,7 @@ mod temporal;
 mod trigonometry;
 
 pub(super) use self::nan::NanFunction;
+pub(super) use self::strings::StringFunction;
 
 use super::*;
 use polars_core::prelude::*;
@@ -64,39 +65,6 @@ pub enum FunctionExpr {
     Nan(NanFunction),
 }
 
-#[cfg(feature = "strings")]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-#[derive(Clone, PartialEq, Debug, Eq, Hash)]
-pub enum StringFunction {
-    Contains {
-        pat: String,
-        literal: bool,
-    },
-    StartsWith(String),
-    EndsWith(String),
-    Extract {
-        pat: String,
-        group_index: usize,
-    },
-    #[cfg(feature = "string_justify")]
-    Zfill(usize),
-    #[cfg(feature = "string_justify")]
-    LJust {
-        width: usize,
-        fillchar: char,
-    },
-    #[cfg(feature = "string_justify")]
-    RJust {
-        width: usize,
-        fillchar: char,
-    },
-    ExtractAll(String),
-    CountMatch(String),
-    #[cfg(feature = "temporal")]
-    Strptime(StrpTimeOptions),
-    #[cfg(feature = "concat_str")]
-    Concat(String),
-}
 
 #[cfg(feature = "trigonometry")]
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
